Conflicts:
	RockWeb/Blocks/Crm/GroupTypes.ascx.cs

+++ RockWeb/Blocks/Crm/GroupTypes.ascx.cs
@@ -7,15 +7,59 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Web;
 using System.Web.UI;
 using Rock;
 using Rock.Constants;
+using Rock.Core;
 using Rock.Crm;
 using Rock.Web.UI;
 using Rock.Web.UI.Controls;
 
 public partial class GroupTypes : RockBlock
 {
+    #region Child Grid Dictionarys
+
+    /// <summary>
+    /// Gets the child group types dictionary.
+    /// </summary>
+    /// <returns></returns>
+    private Dictionary<int, string> ChildGroupTypesDictionary
+    {
+        get
+        {
+            Dictionary<int, string> childGroupTypesDictionary = ViewState["ChildGroupTypesDictionary"] as Dictionary<int, string>;
+            return childGroupTypesDictionary;
+        }
+
+        set
+        {
+            ViewState["ChildGroupTypesDictionary"] = value;
+        }
+    }
+
+    /// <summary>
+    /// Gets or sets the location types dictionary.
+    /// </summary>
+    /// <value>
+    /// The location types dictionary.
+    /// </value>
+    private Dictionary<int, string> LocationTypesDictionary
+    {
+        get
+        {
+            Dictionary<int, string> locationTypesDictionary = ViewState["LocationTypesDictionary"] as Dictionary<int, string>;
+            return locationTypesDictionary;
+        }
+
+        set
+        {
+            ViewState["LocationTypesDictionary"] = value;
+        }
+    }
+
+    #endregion
+
     #region Control Methods
 
     /// <summary>
@@ -37,6 +81,13 @@ public partial class GroupTypes : RockBlock
             gChildGroupTypes.Actions.IsAddEnabled = true;
             gChildGroupTypes.Actions.AddClick += gChildGroupTypes_Add;
             gChildGroupTypes.GridRebind += gChildGroupTypes_GridRebind;
+            gChildGroupTypes.EmptyDataText = Server.HtmlEncode(None.Text);
+
+            gLocationTypes.DataKeyNames = new string[] { "key" };
+            gLocationTypes.Actions.IsAddEnabled = true;
+            gLocationTypes.Actions.AddClick += gLocationTypes_Add;
+            gLocationTypes.GridRebind += gLocationTypes_GridRebind;
+            gLocationTypes.EmptyDataText = Server.HtmlEncode( None.Text );
         }
     }
 
@@ -46,7 +97,8 @@ public partial class GroupTypes : RockBlock
     /// <param name="e">The <see cref="T:System.EventArgs" /> object that contains the event data.</param>
     protected override void OnLoad( EventArgs e )
     {
-        nbMessage.Visible = false;
+        nbGridWarning.Visible = false;
+        nbWarning.Visible = false;
 
         if ( CurrentPage.IsAuthorized( "Configure", CurrentPerson ) )
         {
@@ -59,15 +111,15 @@ public partial class GroupTypes : RockBlock
         else
         {
             gGroupType.Visible = false;
-            nbMessage.Text = WarningMessage.NotAuthorizedToEdit( GroupType.FriendlyTypeName );
-            nbMessage.Visible = true;
+            nbWarning.Text = WarningMessage.NotAuthorizedToEdit( GroupType.FriendlyTypeName );
+            nbWarning.Visible = true;
         }
 
         base.OnLoad( e );
     }
     #endregion
 
-    #region Grid Events
+    #region Grid Events (main grid)
 
     /// <summary>
     /// Handles the Add event of the gGroupType control.
@@ -97,7 +149,16 @@ public partial class GroupTypes : RockBlock
     protected void gGroupType_Delete( object sender, RowEventArgs e )
     {
         GroupTypeService groupTypeService = new GroupTypeService();
-        GroupType groupType = groupTypeService.Get( (int)gGroupType.DataKeys[e.RowIndex]["id"] );
+        int groupTypeId = (int)gGroupType.DataKeys[e.RowIndex]["id"];
+        if ( groupTypeService.IsChildGroupType( groupTypeId ) )
+        {
+            string errorMessage = "This group type is assigned as a child group type.";
+            nbGridWarning.Text = errorMessage;
+            nbGridWarning.Visible = true;
+            return;
+        }
+
+        GroupType groupType = groupTypeService.Get( groupTypeId  );
         if ( CurrentBlock != null )
         {
             groupTypeService.Delete( groupType, CurrentPersonId );
@@ -117,17 +178,64 @@ public partial class GroupTypes : RockBlock
         BindGrid();
     }
 
+    #endregion
 
+    #region Child GroupType Grid and Picker
+
+    /// <summary>
+    /// Handles the Add event of the gChildGroupTypes control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
     protected void gChildGroupTypes_Add( object sender, EventArgs e )
     {
-        throw new NotImplementedException();
+        GroupTypeService groupTypeService = new GroupTypeService();
+        int currentGroupTypeId = int.Parse( hfGroupTypeId.Value );
+
+        // populate dropdown with all grouptypes that aren't already childgroups or the current grouptype
+        var qry = from gt in groupTypeService.Queryable()
+                  where gt.Id != currentGroupTypeId
+                  where !( from cgt in ChildGroupTypesDictionary.Keys
+                           select cgt ).Contains( gt.Id )
+                  select gt;
+
+        List<GroupType> list = qry.ToList();
+        if ( list.Count == 0 )
+        {
+            list.Add( new GroupType { Id = None.Id, Name = None.Text } );
+            btnAddChildGroupType.Enabled = false;
+            btnAddChildGroupType.CssClass = "btn primary disabled";
+        }
+        else
+        {
+            btnAddChildGroupType.Enabled = true;
+            btnAddChildGroupType.CssClass = "btn primary";
+        }
+
+        ddlChildGroupType.DataSource = list;
+        ddlChildGroupType.DataBind();
+        pnlChildGroupTypePicker.Visible = true;
+        pnlDetails.Visible = false;
+        pnlList.Visible = false;
     }
 
+    /// <summary>
+    /// Handles the Delete event of the gChildGroupTypes control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="RowEventArgs" /> instance containing the event data.</param>
     protected void gChildGroupTypes_Delete( object sender, RowEventArgs e )
     {
-        throw new NotImplementedException();
+        int childGroupTypeId = (int)gChildGroupTypes.DataKeys[e.RowIndex]["key"];
+        ChildGroupTypesDictionary.Remove( childGroupTypeId );
+        BindChildGroupTypesGrid();
     }
 
+    /// <summary>
+    /// Handles the GridRebind event of the gChildGroupTypes control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
     protected void gChildGroupTypes_GridRebind( object sender, EventArgs e )
     {
         BindChildGroupTypesGrid();
@@ -138,8 +246,134 @@ public partial class GroupTypes : RockBlock
     /// </summary>
     private void BindChildGroupTypesGrid()
     {
-        gChildGroupTypes.DataSource = ViewState["ChildGroupTypes"];
+        gChildGroupTypes.DataSource = ChildGroupTypesDictionary.OrderBy( a => a.Value );
         gChildGroupTypes.DataBind();
+    }
+
+    /// <summary>
+    /// Handles the Click event of the btnAddChildGroupType control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
+    protected void btnAddChildGroupType_Click( object sender, EventArgs e )
+    {
+        ChildGroupTypesDictionary.Add( int.Parse( ddlChildGroupType.SelectedValue ), ddlChildGroupType.SelectedItem.Text );
+
+        pnlChildGroupTypePicker.Visible = false;
+        pnlDetails.Visible = true;
+        pnlList.Visible = false;
+
+        BindChildGroupTypesGrid();
+    }
+
+    /// <summary>
+    /// Handles the Click event of the btnCancelAddChildGroupType control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
+    protected void btnCancelAddChildGroupType_Click( object sender, EventArgs e )
+    {
+        pnlChildGroupTypePicker.Visible = false;
+        pnlDetails.Visible = true;
+        pnlList.Visible = false;
+    }
+
+    #endregion
+
+    #region LocationType Grid and Picker
+
+    /// <summary>
+    /// Handles the Add event of the gLocationTypes control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
+    protected void gLocationTypes_Add( object sender, EventArgs e )
+    {
+        DefinedValueService definedValueService = new DefinedValueService();
+
+        // populate dropdown with all locationtypes that aren't already locationtypes
+        var qry = from dlt in definedValueService.GetByDefinedTypeGuid( Rock.SystemGuid.DefinedType.LOCATION_LOCATION_TYPE )
+                  where !( from lt in LocationTypesDictionary.Keys
+                           select lt ).Contains( dlt.Id )
+                  select dlt;
+
+        List<DefinedValue> list = qry.ToList();
+        if ( list.Count == 0 )
+        {
+            list.Add( new DefinedValue { Id = None.Id, Name = None.Text } );
+            btnAddLocationType.Enabled = false;
+            btnAddLocationType.CssClass = "btn primary disabled";
+        }
+        else
+        {
+            btnAddLocationType.Enabled = true;
+            btnAddLocationType.CssClass = "btn primary";
+        }
+
+        ddlLocationType.DataSource = list;
+        ddlLocationType.DataBind();
+        pnlLocationTypePicker.Visible = true;
+        pnlDetails.Visible = false;
+        pnlList.Visible = false;
+    }
+
+    /// <summary>
+    /// Handles the Delete event of the gLocationTypes control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="RowEventArgs" /> instance containing the event data.</param>
+    protected void gLocationTypes_Delete( object sender, RowEventArgs e )
+    {
+        int locationTypeId = (int)gLocationTypes.DataKeys[e.RowIndex]["key"];
+        LocationTypesDictionary.Remove( locationTypeId );
+        BindLocationTypesGrid();
+    }
+
+    /// <summary>
+    /// Handles the GridRebind event of the gLocationTypes control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
+    protected void gLocationTypes_GridRebind( object sender, EventArgs e )
+    {
+        BindLocationTypesGrid();
+    }
+
+    /// <summary>
+    /// Binds the location types grid.
+    /// </summary>
+    private void BindLocationTypesGrid()
+    {
+        gLocationTypes.DataSource = LocationTypesDictionary.OrderBy( a => a.Value );
+        gLocationTypes.DataBind();
+    }
+
+    /// <summary>
+    /// Handles the Click event of the btnAddLocationType control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
+    protected void btnAddLocationType_Click( object sender, EventArgs e )
+    {
+        LocationTypesDictionary.Add( int.Parse( ddlLocationType.SelectedValue ), ddlLocationType.SelectedItem.Text );
+
+        pnlLocationTypePicker.Visible = false;
+        pnlDetails.Visible = true;
+        pnlList.Visible = false;
+
+        BindLocationTypesGrid();
+    }
+
+    /// <summary>
+    /// Handles the Click event of the btnCancelAddLocationType control.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="e">The <see cref="EventArgs" /> instance containing the event data.</param>
+    protected void btnCancelAddLocationType_Click( object sender, EventArgs e )
+    {
+        pnlLocationTypePicker.Visible = false;
+        pnlDetails.Visible = true;
+        pnlList.Visible = false;
     }
 
     #endregion
@@ -191,11 +425,34 @@ public partial class GroupTypes : RockBlock
 
         groupType.Description = tbDescription.Text;
 
+        groupType.ChildGroupTypes = new List<GroupType>();
+        groupType.ChildGroupTypes.Clear();
+        foreach ( var item in ChildGroupTypesDictionary )
+        {
+            var childGroupType = groupTypeService.Get( item.Key );
+            if ( childGroupType != null )
+            {
+                groupType.ChildGroupTypes.Add( childGroupType );
+            }
+        }
+
+        DefinedValueService definedValueService = new DefinedValueService();
+        
+        groupType.LocationTypes = new List<GroupTypeLocationType>();
+        groupType.LocationTypes.Clear();
+        foreach ( var item in LocationTypesDictionary )
+        {
+            var locationType = definedValueService.Get( item.Key );
+            if ( locationType != null )
+            {
+                groupType.LocationTypes.Add( new GroupTypeLocationType { LocationTypeId = locationType.Id } );
+            }
+        }
+
         // check for duplicates
         if ( groupTypeService.Queryable().Count( a => a.Name.Equals( groupType.Name, StringComparison.OrdinalIgnoreCase ) && !a.Id.Equals( groupType.Id ) ) > 0 )
         {
-            nbMessage.Text = WarningMessage.DuplicateFoundMessage( "name", GroupType.FriendlyTypeName );
-            nbMessage.Visible = true;
+            tbName.ShowErrorMessage( WarningMessage.DuplicateFoundMessage( "name", GroupType.FriendlyTypeName ) );
             return;
         }
 
@@ -260,6 +517,8 @@ public partial class GroupTypes : RockBlock
         GroupTypeService groupTypeService = new GroupTypeService();
         GroupType groupType = groupTypeService.Get( groupTypeId );
         bool readOnly = false;
+        ChildGroupTypesDictionary = new Dictionary<int, string>();
+        LocationTypesDictionary = new Dictionary<int, string>();
 
         if ( groupType != null )
         {
@@ -267,9 +526,8 @@ public partial class GroupTypes : RockBlock
             tbName.Text = groupType.Name;
             tbDescription.Text = groupType.Description;
             ddlDefaultGroupRole.SelectedValue = ( groupType.DefaultGroupRoleId ?? None.Id ).ToString();
-            var list = new Dictionary<int, string>();
-            groupType.ChildGroupTypes.ToList().ForEach( a => list.Add( a.Id, a.Name ) );
-            ViewState["ChildGroupTypes"] = list;
+            groupType.ChildGroupTypes.ToList().ForEach( a => ChildGroupTypesDictionary.Add( a.Id, a.Name ) );
+            groupType.LocationTypes.ToList().ForEach( a => LocationTypesDictionary.Add( a.LocationTypeId, a.LocationType.Name ) );
             readOnly = groupType.IsSystem;
 
             if ( groupType.IsSystem )
@@ -290,7 +548,6 @@ public partial class GroupTypes : RockBlock
             hfGroupTypeId.Value = 0.ToString();
             tbName.Text = string.Empty;
             tbDescription.Text = string.Empty;
-            ViewState["ChildGroupTypes"] = new Dictionary<int, string>();
         }
 
         iconIsSystem.Visible = readOnly;
@@ -300,6 +557,7 @@ public partial class GroupTypes : RockBlock
         btnSave.Visible = !readOnly;
 
         BindChildGroupTypesGrid();
+        BindLocationTypesGrid();
     }
 
     #endregion
